fix deliberation type
