+++ ts-packages/web/src/app/spaces/[id]/deliberation/_components/charts/models/parsed.tsx
@@ -1,0 +1,24 @@
+import {
+  CheckboxQuestion,
+  DropdownQuestion,
+  LinearScaleQuestion,
+  MultipleChoiceQuestion,
+  SingleChoiceQuestion,
+} from '@/lib/api/models/survey';
+
+type ParsedOption = {
+  label: string;
+  count: number;
+  ratio: number;
+};
+
+export type ParsedResult = {
+  question:
+    | SingleChoiceQuestion
+    | MultipleChoiceQuestion
+    | CheckboxQuestion
+    | DropdownQuestion
+    | LinearScaleQuestion;
+  totalParticipants: number;
+  options: ParsedOption[];
+};
